feat(vote): seprate vote view to reuse
